feat(workspace-cap): migrate embeddings + embedding_cache (slice 6 of N)
ADR-152 section 3 F3.4 slice 6. Verbatim port of embeddings.rs + embedding_cache.rs
into dasclaw_workspace_cap, with NearAiEmbeddings split out to ironclaw
because it depends on the business-layer SessionManager.

Closes #702
Refs #674, #701
