+++ crates/dasclaw_workspace_cap/Cargo.toml
@@ -29,6 +29,18 @@ chrono = { version = "0.4", features = ["serde"] }
 uuid = { version = "1", features = ["v4", "serde"] }
 serde_json = "1"
 
+# Embeddings + cache (slice 6/N: embeddings, embedding_cache). Mirrors
+# ironclaw's existing versions; reqwest stays on the rustls-tls-native-roots
+# minimal feature set (only JSON POST is needed for embedding providers).
+async-trait = "0.1"
+reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls-native-roots"] }
+lru = "0.16.3"
+sha2 = "0.10"
+tracing = "0.1"
+# OpenAI embedding 429 handling reuses the shared `parse_retry_after` helper
+# from the llm provider crate (no behavior change vs ironclaw).
+dasclaw_llm_provider = { path = "../dasclaw_llm_provider" }
+
 # PostgreSQL persistence (slice 5/N: repository). Optional, gated by `postgres`
 # feature so default builds stay dependency-light. Mirrors ironclaw's existing
 # `postgres` feature set (see desktop-client/ironclaw/Cargo.toml).
@@ -50,3 +62,4 @@ postgres = [
 
 [dev-dependencies]
 tempfile = "3"
+tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
