+++ crates/dasclaw_workspace_cap/src/embedding_cache.rs
@@ -12,7 +12,11 @@ use async_trait::async_trait;
 use lru::LruCache;
 use sha2::{Digest, Sha256};
 
-use crate::workspace::embeddings::{EmbeddingError, EmbeddingProvider};
+use crate::embeddings::{EmbeddingError, EmbeddingProvider};
+
+/// Default maximum number of cached embeddings (mirrors ironclaw's
+/// `config::embeddings::DEFAULT_EMBEDDING_CACHE_SIZE`).
+pub const DEFAULT_EMBEDDING_CACHE_SIZE: usize = 10_000;
 
 /// Configuration for the embedding cache.
 #[derive(Debug, Clone)]
@@ -28,7 +32,7 @@ pub struct EmbeddingCacheConfig {
 impl Default for EmbeddingCacheConfig {
     fn default() -> Self {
         Self {
-            max_entries: crate::config::DEFAULT_EMBEDDING_CACHE_SIZE,
+            max_entries: DEFAULT_EMBEDDING_CACHE_SIZE,
         }
     }
 }
